requirements.txt
# requirements.txt file
Django==4.1.6
django-environ==0.9.0
mysqlclient==2.1.1
PyMySQL==1.0.2